More tests, fuzz tests
